+++ paymentchannel/src/Commitment.cpp
@@ -17,7 +17,7 @@ namespace paymentchannel {
         : _value(0) {
     }
 
-    Commitment::Commitment(int64_t value, const Coin::PublicKey & payorPk, const Coin::PublicKey & payeePk, uint32_t lockTime)
+    Commitment::Commitment(int64_t value, const Coin::PublicKey & payorPk, const Coin::PublicKey & payeePk, uint16_t lockTime)
         : _value(value)
         , _payorPk(payorPk)
         , _payeePk(payeePk)
@@ -74,11 +74,11 @@ namespace paymentchannel {
         _payeePk = payeePk;
     }
 
-    void Commitment::setLockTime(uint32_t lockTime) {
+    void Commitment::setLockTime(uint16_t lockTime) {
         _lockTime = lockTime;
     }
 
-    uint32_t Commitment::lockTime() const {
+    uint16_t Commitment::lockTime() const {
         return _lockTime;
     }
 }
